re-enable ClarityInstance tests by moving them to clarity_vm, add trait for clarity::TransactionConnection \w auto-impls
