+++ src/clarity_vm/tests/large_contract.rs
@@ -16,8 +16,7 @@
 
 use crate::types::chainstate::BlockHeaderHash;
 use crate::types::chainstate::StacksBlockId;
-use crate::types::proof::ClarityMarfTrieId;
-use chainstate::stacks::index::storage::TrieFileStorage;
+use chainstate::stacks::index::ClarityMarfTrieId;
 use clarity_vm::clarity::{ClarityInstance, Error as ClarityError};
 use util::hash::hex_bytes;
 use vm::ast;
@@ -26,19 +25,282 @@ use vm::contracts::Contract;
 use vm::costs::ExecutionCost;
 use vm::database::ClarityDatabase;
 use vm::errors::{CheckErrors, Error, RuntimeErrorType};
-use vm::execute as vm_execute;
 use vm::representations::SymbolicExpression;
-use vm::tests::{
-    execute, symbols_from_values, with_marfed_environment, with_memory_environment,
-    TEST_BURN_STATE_DB, TEST_HEADER_DB,
-};
+use vm::test_util::*;
 use vm::types::{
     OptionalData, PrincipalData, QualifiedContractIdentifier, ResponseData, StandardPrincipalData,
     TypeSignature, Value,
 };
 
 use crate::clarity_vm::database::marf::MarfedKV;
-use vm::database::MemoryBackingStore;
+use vm::clarity::TransactionConnection;
+
+fn test_block_headers(n: u8) -> StacksBlockId {
+    StacksBlockId([n as u8; 32])
+}
+
+const SIMPLE_TOKENS: &str = "(define-map tokens { account: principal } { balance: uint })
+         (define-read-only (my-get-token-balance (account principal))
+            (default-to u0 (get balance (map-get? tokens (tuple (account account))))))
+         (define-read-only (explode (account principal))
+             (map-delete tokens (tuple (account account))))
+         (define-private (token-credit! (account principal) (amount uint))
+            (if (<= amount u0)
+                (err \"must be positive\")
+                (let ((current-amount (my-get-token-balance account)))
+                  (begin
+                    (map-set tokens (tuple (account account))
+                                       (tuple (balance (+ amount current-amount))))
+                    (ok 0)))))
+         (define-public (token-transfer (to principal) (amount uint))
+          (let ((balance (my-get-token-balance tx-sender)))
+             (if (or (> amount balance) (<= amount u0))
+                 (err \"not enough balance\")
+                 (begin
+                   (map-set tokens (tuple (account tx-sender))
+                                      (tuple (balance (- balance amount))))
+                   (token-credit! to amount)))))
+         (define-public (faucet)
+           (let ((original-sender tx-sender))
+             (as-contract (print (token-transfer (print original-sender) u1)))))                     
+         (define-public (mint-after (block-to-release uint))
+           (if (>= block-height block-to-release)
+               (faucet)
+               (err \"must be in the future\")))
+         (begin (token-credit! 'SZ2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR u10000)
+                (token-credit! 'SM2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQVX8X0G u200)
+                (token-credit! .tokens u4))";
+
+#[test]
+fn test_simple_token_system() {
+    let mut clarity = ClarityInstance::new(false, MarfedKV::temporary());
+    let p1 = PrincipalData::from(
+        PrincipalData::parse_standard_principal("SZ2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR")
+            .unwrap(),
+    );
+    let p2 = PrincipalData::from(
+        PrincipalData::parse_standard_principal("SM2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQVX8X0G")
+            .unwrap(),
+    );
+    let contract_identifier = QualifiedContractIdentifier::local("tokens").unwrap();
+
+    {
+        let mut block = clarity.begin_test_genesis_block(
+            &StacksBlockId::sentinel(),
+            &test_block_headers(0),
+            &TEST_HEADER_DB,
+            &TEST_BURN_STATE_DB,
+        );
+
+        let tokens_contract = SIMPLE_TOKENS;
+
+        let contract_ast = ast::build_ast(&contract_identifier, tokens_contract, &mut ()).unwrap();
+
+        block.as_transaction(|tx| {
+            tx.initialize_smart_contract(
+                &contract_identifier,
+                &contract_ast,
+                tokens_contract,
+                |_, _| false,
+            )
+            .unwrap()
+        });
+
+        assert!(!is_committed(
+            &block
+                .as_transaction(|tx| tx.run_contract_call(
+                    &p2,
+                    &contract_identifier,
+                    "token-transfer",
+                    &[p1.clone().into(), Value::UInt(210)],
+                    |_, _| false
+                ))
+                .unwrap()
+                .0
+        ));
+        assert!(is_committed(
+            &block
+                .as_transaction(|tx| tx.run_contract_call(
+                    &p1,
+                    &contract_identifier,
+                    "token-transfer",
+                    &[p2.clone().into(), Value::UInt(9000)],
+                    |_, _| false
+                ))
+                .unwrap()
+                .0
+        ));
+
+        assert!(!is_committed(
+            &block
+                .as_transaction(|tx| tx.run_contract_call(
+                    &p1,
+                    &contract_identifier,
+                    "token-transfer",
+                    &[p2.clone().into(), Value::UInt(1001)],
+                    |_, _| false
+                ))
+                .unwrap()
+                .0
+        ));
+        assert!(is_committed(
+            & // send to self!
+            block.as_transaction(|tx| tx.run_contract_call(&p1, &contract_identifier, "token-transfer",
+                                    &[p1.clone().into(), Value::UInt(1000)], |_, _| false)).unwrap().0
+        ));
+
+        assert_eq!(
+            block
+                .as_transaction(|tx| tx.eval_read_only(
+                    &contract_identifier,
+                    "(my-get-token-balance 'SZ2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR)"
+                ))
+                .unwrap(),
+            Value::UInt(1000)
+        );
+        assert_eq!(
+            block
+                .as_transaction(|tx| tx.eval_read_only(
+                    &contract_identifier,
+                    "(my-get-token-balance 'SM2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQVX8X0G)"
+                ))
+                .unwrap(),
+            Value::UInt(9200)
+        );
+
+        assert!(is_committed(
+            &block
+                .as_transaction(|tx| tx.run_contract_call(
+                    &p1,
+                    &contract_identifier,
+                    "faucet",
+                    &[],
+                    |_, _| false
+                ))
+                .unwrap()
+                .0
+        ));
+
+        assert!(is_committed(
+            &block
+                .as_transaction(|tx| tx.run_contract_call(
+                    &p1,
+                    &contract_identifier,
+                    "faucet",
+                    &[],
+                    |_, _| false
+                ))
+                .unwrap()
+                .0
+        ));
+
+        assert!(is_committed(
+            &block
+                .as_transaction(|tx| tx.run_contract_call(
+                    &p1,
+                    &contract_identifier,
+                    "faucet",
+                    &[],
+                    |_, _| false
+                ))
+                .unwrap()
+                .0
+        ));
+
+        assert_eq!(
+            block
+                .as_transaction(|tx| tx.eval_read_only(
+                    &contract_identifier,
+                    "(my-get-token-balance 'SZ2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR)"
+                ))
+                .unwrap(),
+            Value::UInt(1003)
+        );
+
+        assert!(!is_committed(
+            &block
+                .as_transaction(|tx| tx.run_contract_call(
+                    &p1,
+                    &contract_identifier,
+                    "mint-after",
+                    &[Value::UInt(25)],
+                    |_, _| false
+                ))
+                .unwrap()
+                .0
+        ));
+        block.commit_block();
+    }
+
+    for i in 0..25 {
+        {
+            let block = clarity.begin_block(
+                &test_block_headers(i),
+                &test_block_headers(i + 1),
+                &TEST_HEADER_DB,
+                &TEST_BURN_STATE_DB,
+            );
+            block.commit_block();
+        }
+    }
+
+    {
+        let mut block = clarity.begin_block(
+            &test_block_headers(25),
+            &test_block_headers(26),
+            &TEST_HEADER_DB,
+            &TEST_BURN_STATE_DB,
+        );
+        assert!(is_committed(
+            &block
+                .as_transaction(|tx| tx.run_contract_call(
+                    &p1,
+                    &contract_identifier,
+                    "mint-after",
+                    &[Value::UInt(25)],
+                    |_, _| false
+                ))
+                .unwrap()
+                .0
+        ));
+
+        assert!(!is_committed(
+            &block
+                .as_transaction(|tx| tx.run_contract_call(
+                    &p1,
+                    &contract_identifier,
+                    "faucet",
+                    &[],
+                    |_, _| false
+                ))
+                .unwrap()
+                .0
+        ));
+
+        assert_eq!(
+            block
+                .as_transaction(|tx| tx.eval_read_only(
+                    &contract_identifier,
+                    "(my-get-token-balance 'SZ2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR)"
+                ))
+                .unwrap(),
+            Value::UInt(1004)
+        );
+        assert_eq!(
+            block
+                .as_transaction(|tx| tx.run_contract_call(
+                    &p1,
+                    &contract_identifier,
+                    "my-get-token-balance",
+                    &[p1.clone().into()],
+                    |_, _| false
+                ))
+                .unwrap()
+                .0,
+            Value::UInt(1004)
+        );
+    }
+}
 
 /*
  * This test exhibits memory inflation --
