+++ src/clarity_vm/tests/mod.rs
@@ -14,4 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+pub mod ast;
+pub mod costs;
 pub mod epoch_switch;
+pub mod forking;
+pub mod large_contract;
