+++ src/vm/ast/mod.rs
@@ -70,66 +70,13 @@ pub fn build_ast<T: CostTracker>(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::types::chainstate::StacksBlockId;
+mod test {
     use std::collections::HashMap;
-    use vm::costs::*;
-    use vm::database::*;
+
+    use vm::ast::build_ast;
+    use vm::costs::LimitedCostTracker;
     use vm::representations::depth_traverse;
-    use vm::tests::{TEST_BURN_STATE_DB, TEST_HEADER_DB};
-
-    fn dependency_edge_counting_runtime(iters: usize) -> u64 {
-        // let mut progn = "(define-private (a0) 1)".to_string();
-        // for i in 1..iters {
-        //     progn.push_str(&format!("\n(define-private (a{}) (begin", i));
-        //     for x in 0..i {
-        //         progn.push_str(&format!(" (a{}) ", x));
-        //     }
-        //     progn.push_str("))");
-        // }
-
-        // let marf = MarfedKV::temporary();
-        // let mut clarity_instance = ClarityInstance::new(false, marf);
-
-        // clarity_instance
-        //     .begin_test_genesis_block(
-        //         &StacksBlockId::sentinel(),
-        //         &StacksBlockId([0 as u8; 32]),
-        //         &TEST_HEADER_DB,
-        //         &TEST_BURN_STATE_DB,
-        //     )
-        //     .commit_block();
-
-        // let mut cost_track = clarity_instance
-        //     .begin_block(
-        //         &StacksBlockId([0 as u8; 32]),
-        //         &StacksBlockId([1 as u8; 32]),
-        //         &TEST_HEADER_DB,
-        //         &TEST_BURN_STATE_DB,
-        //     )
-        //     .commit_block();
-
-        // build_ast(
-        //     &QualifiedContractIdentifier::transient(),
-        //     &progn,
-        //     &mut cost_track,
-        // )
-        // .unwrap();
-
-        // cost_track.get_total().runtime
-        1
-    }
-
-    #[test]
-    fn test_edge_counting_runtime() {
-        let ratio_4_8 = dependency_edge_counting_runtime(8) / dependency_edge_counting_runtime(4);
-        let ratio_8_16 = dependency_edge_counting_runtime(16) / dependency_edge_counting_runtime(8);
-
-        // this really is just testing for the non-linearity
-        //   in the runtime cost assessment (because the edge count in the dependency graph is going up O(n^2)).
-        assert!(ratio_8_16 > ratio_4_8);
-    }
+    use vm::types::QualifiedContractIdentifier;
 
     #[test]
     fn test_expression_identification_tuples() {
