+++ src/vm/tests/contracts.rs
@@ -27,8 +27,8 @@ use vm::errors::{CheckErrors, Error, RuntimeErrorType};
 use vm::execute as vm_execute;
 use vm::representations::SymbolicExpression;
 use vm::tests::{
-    execute, symbols_from_values, with_marfed_environment, with_memory_environment,
-    TEST_BURN_STATE_DB, TEST_HEADER_DB,
+    execute, is_committed, is_err_code_i128 as is_err_code, symbols_from_values,
+    with_marfed_environment, with_memory_environment, TEST_BURN_STATE_DB, TEST_HEADER_DB,
 };
 use vm::types::{
     OptionalData, PrincipalData, QualifiedContractIdentifier, ResponseData, StandardPrincipalData,
@@ -162,255 +162,6 @@ fn test_get_block_info_eval() {
         }
     }
 }
-
-fn is_committed(v: &Value) -> bool {
-    match v {
-        Value::Response(ref data) => data.committed,
-        _ => false,
-    }
-}
-
-fn is_err_code(v: &Value, e: i128) -> bool {
-    match v {
-        Value::Response(ref data) => !data.committed && *data.data == Value::Int(e),
-        _ => false,
-    }
-}
-
-fn test_block_headers(n: u8) -> StacksBlockId {
-    StacksBlockId([n as u8; 32])
-}
-
-// #[test]
-// fn test_simple_token_system() {
-//     let mut clarity = ClarityInstance::new(false, MarfedKV::temporary());
-//     let p1 = PrincipalData::from(
-//         PrincipalData::parse_standard_principal("SZ2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR")
-//             .unwrap(),
-//     );
-//     let p2 = PrincipalData::from(
-//         PrincipalData::parse_standard_principal("SM2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQVX8X0G")
-//             .unwrap(),
-//     );
-//     let contract_identifier = QualifiedContractIdentifier::local("tokens").unwrap();
-
-//     {
-//         let mut block = clarity.begin_test_genesis_block(
-//             &StacksBlockId::sentinel(),
-//             &test_block_headers(0),
-//             &TEST_HEADER_DB,
-//             &TEST_BURN_STATE_DB,
-//         );
-
-//         let tokens_contract = SIMPLE_TOKENS;
-
-//         let contract_ast = ast::build_ast(&contract_identifier, tokens_contract, &mut ()).unwrap();
-
-//         block.as_transaction(|tx| {
-//             tx.initialize_smart_contract(
-//                 &contract_identifier,
-//                 &contract_ast,
-//                 tokens_contract,
-//                 |_, _| false,
-//             )
-//             .unwrap()
-//         });
-
-//         assert!(!is_committed(
-//             &block
-//                 .as_transaction(|tx| tx.run_contract_call(
-//                     &p2,
-//                     &contract_identifier,
-//                     "token-transfer",
-//                     &[p1.clone().into(), Value::UInt(210)],
-//                     |_, _| false
-//                 ))
-//                 .unwrap()
-//                 .0
-//         ));
-//         assert!(is_committed(
-//             &block
-//                 .as_transaction(|tx| tx.run_contract_call(
-//                     &p1,
-//                     &contract_identifier,
-//                     "token-transfer",
-//                     &[p2.clone().into(), Value::UInt(9000)],
-//                     |_, _| false
-//                 ))
-//                 .unwrap()
-//                 .0
-//         ));
-
-//         assert!(!is_committed(
-//             &block
-//                 .as_transaction(|tx| tx.run_contract_call(
-//                     &p1,
-//                     &contract_identifier,
-//                     "token-transfer",
-//                     &[p2.clone().into(), Value::UInt(1001)],
-//                     |_, _| false
-//                 ))
-//                 .unwrap()
-//                 .0
-//         ));
-//         assert!(is_committed(
-//             & // send to self!
-//             block.as_transaction(|tx| tx.run_contract_call(&p1, &contract_identifier, "token-transfer",
-//                                     &[p1.clone().into(), Value::UInt(1000)], |_, _| false)).unwrap().0
-//         ));
-
-//         assert_eq!(
-//             block
-//                 .as_transaction(|tx| tx.eval_read_only(
-//                     &contract_identifier,
-//                     "(my-get-token-balance 'SZ2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR)"
-//                 ))
-//                 .unwrap(),
-//             Value::UInt(1000)
-//         );
-//         assert_eq!(
-//             block
-//                 .as_transaction(|tx| tx.eval_read_only(
-//                     &contract_identifier,
-//                     "(my-get-token-balance 'SM2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQVX8X0G)"
-//                 ))
-//                 .unwrap(),
-//             Value::UInt(9200)
-//         );
-
-//         assert!(is_committed(
-//             &block
-//                 .as_transaction(|tx| tx.run_contract_call(
-//                     &p1,
-//                     &contract_identifier,
-//                     "faucet",
-//                     &[],
-//                     |_, _| false
-//                 ))
-//                 .unwrap()
-//                 .0
-//         ));
-
-//         assert!(is_committed(
-//             &block
-//                 .as_transaction(|tx| tx.run_contract_call(
-//                     &p1,
-//                     &contract_identifier,
-//                     "faucet",
-//                     &[],
-//                     |_, _| false
-//                 ))
-//                 .unwrap()
-//                 .0
-//         ));
-
-//         assert!(is_committed(
-//             &block
-//                 .as_transaction(|tx| tx.run_contract_call(
-//                     &p1,
-//                     &contract_identifier,
-//                     "faucet",
-//                     &[],
-//                     |_, _| false
-//                 ))
-//                 .unwrap()
-//                 .0
-//         ));
-
-//         assert_eq!(
-//             block
-//                 .as_transaction(|tx| tx.eval_read_only(
-//                     &contract_identifier,
-//                     "(my-get-token-balance 'SZ2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR)"
-//                 ))
-//                 .unwrap(),
-//             Value::UInt(1003)
-//         );
-
-//         assert!(!is_committed(
-//             &block
-//                 .as_transaction(|tx| tx.run_contract_call(
-//                     &p1,
-//                     &contract_identifier,
-//                     "mint-after",
-//                     &[Value::UInt(25)],
-//                     |_, _| false
-//                 ))
-//                 .unwrap()
-//                 .0
-//         ));
-//         block.commit_block();
-//     }
-
-//     for i in 0..25 {
-//         {
-//             let block = clarity.begin_block(
-//                 &test_block_headers(i),
-//                 &test_block_headers(i + 1),
-//                 &TEST_HEADER_DB,
-//                 &TEST_BURN_STATE_DB,
-//             );
-//             block.commit_block();
-//         }
-//     }
-
-//     {
-//         let mut block = clarity.begin_block(
-//             &test_block_headers(25),
-//             &test_block_headers(26),
-//             &TEST_HEADER_DB,
-//             &TEST_BURN_STATE_DB,
-//         );
-//         assert!(is_committed(
-//             &block
-//                 .as_transaction(|tx| tx.run_contract_call(
-//                     &p1,
-//                     &contract_identifier,
-//                     "mint-after",
-//                     &[Value::UInt(25)],
-//                     |_, _| false
-//                 ))
-//                 .unwrap()
-//                 .0
-//         ));
-
-//         assert!(!is_committed(
-//             &block
-//                 .as_transaction(|tx| tx.run_contract_call(
-//                     &p1,
-//                     &contract_identifier,
-//                     "faucet",
-//                     &[],
-//                     |_, _| false
-//                 ))
-//                 .unwrap()
-//                 .0
-//         ));
-
-//         assert_eq!(
-//             block
-//                 .as_transaction(|tx| tx.eval_read_only(
-//                     &contract_identifier,
-//                     "(my-get-token-balance 'SZ2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR)"
-//                 ))
-//                 .unwrap(),
-//             Value::UInt(1004)
-//         );
-//         assert_eq!(
-//             block
-//                 .as_transaction(|tx| tx.run_contract_call(
-//                     &p1,
-//                     &contract_identifier,
-//                     "my-get-token-balance",
-//                     &[p1.clone().into()],
-//                     |_, _| false
-//                 ))
-//                 .unwrap()
-//                 .0,
-//             Value::UInt(1004)
-//         );
-//     }
-// }
 
 fn test_contract_caller(owned_env: &mut OwnedEnvironment) {
     let contract_a = "(define-read-only (get-caller)
