+++ src/vm/tests/mod.rs
@@ -20,8 +20,6 @@ use vm::contracts::Contract;
 use vm::database::ClarityDatabase;
 use vm::database::MemoryBackingStore;
 use vm::errors::Error;
-use vm::execute as vm_execute;
-use vm::execute_on_network as vm_execute_on_network;
 use vm::representations::SymbolicExpression;
 use vm::types::{PrincipalData, ResponseData, Value};
 use vm::StacksEpoch;
@@ -35,22 +33,126 @@ use stacks_common::types::chainstate::{
 };
 use stacks_common::types::{StacksEpochId, PEER_VERSION_EPOCH_2_0};
 
-use super::costs::ExecutionCost;
-use super::database::{BurnStateDB, HeadersDB};
+use vm::{
+    analysis::AnalysisDatabase,
+    clarity::{ClarityConnection, TransactionConnection},
+    contexts::AssetMap,
+    costs::{ExecutionCost, LimitedCostTracker},
+    database::{BurnStateDB, HeadersDB},
+};
 
+use super::events::StacksTransactionEvent;
 pub use super::test_util::*;
 
 mod assets;
 mod contracts;
-// pub mod costs;
 mod datamaps;
 mod defines;
-// mod events;
-// mod forking;
-// mod large_contract;
+mod events;
 mod sequences;
 mod simple_apply_eval;
 mod traits;
+
+// pub struct TestBlockConnection {
+//     store: MemoryBackingStore,
+//     epoch: StacksEpochId,
+// }
+
+// pub struct TestTxConnection<'a> {
+//     store: &'a mut MemoryBackingStore,
+//     epoch: StacksEpochId,
+// }
+
+// impl TestBlockConnection {
+//     pub fn new(epoch: StacksEpochId) -> TestBlockConnection {
+//         TestBlockConnection {
+//             store: MemoryBackingStore::new(),
+//             epoch
+//         }
+//     }
+
+//     pub fn start_transaction_processing<'a>(&'a mut self) -> TestTxConnection<'a> {
+//         TestTxConnection {
+//             store: &mut self.store,
+//             epoch: self.epoch
+//         }
+//     }
+
+//     pub fn as_transaction<F, R>(&mut self, todo: F) -> R
+//     where
+//         F: FnOnce(&mut TestTxConnection) -> R,
+//     {
+//         let mut tx = self.start_transaction_processing();
+//         let r = todo(&mut tx);
+//         tx.commit();
+//         r
+//     }
+// }
+
+// impl <'a> ClarityConnection for TestTxConnection <'a> {
+//     fn with_clarity_db_readonly_owned<F, R>(&mut self, to_do: F) -> R
+//     where
+//         F: FnOnce(ClarityDatabase) -> (R, ClarityDatabase) {
+//         let db = self.store.as_clarity_db();
+//         let (result, _) = to_do(db);
+//         result
+//     }
+
+//     fn with_analysis_db_readonly<F, R>(&mut self, to_do: F) -> R
+//     where
+//         F: FnOnce(&mut AnalysisDatabase) -> R {
+//         let mut db = self.store.as_analysis_db();
+//         to_do(&mut db)
+//     }
+
+//     fn get_epoch(&self) -> StacksEpochId {
+//         self.epoch
+//     }
+// }
+
+// impl <'a> TransactionConnection for TestTxConnection <'a> {
+//     fn with_abort_callback<F, A, R, E>(
+//         &mut self,
+//         to_do: F,
+//         abort_call_back: A,
+//     ) -> Result<(R, AssetMap, Vec<StacksTransactionEvent>, bool), E>
+//     where
+//         A: FnOnce(&AssetMap, &mut ClarityDatabase) -> bool,
+//         F: FnOnce(
+//             &mut OwnedEnvironment,
+//         ) -> Result<(R, AssetMap, Vec<StacksTransactionEvent>), E>
+//     {
+//         let mut db = self.store.as_clarity_db();
+//         db.begin();
+//         let mut owned_env = OwnedEnvironment::new(db);
+//         let result = to_do(&mut owned_env);
+//         let (mut db, cost_track) = owned_env.destruct().unwrap();
+//         match result {
+//             Ok((value, asset_map, events)) => {
+//                 let aborted = abort_call_back(&asset_map, &mut db);
+//                 if aborted {
+//                     db.roll_back();
+//                 } else {
+//                     db.commit();
+//                 }
+//                 Ok((value, asset_map, events, aborted))
+//             }
+//             Err(e) => {
+//                 db.roll_back();
+//                 Err(e)
+//             }
+//         }
+//     }
+
+//     fn with_analysis_db<F, R>(&mut self, to_do: F) -> R
+//     where
+//         F: FnOnce(&mut AnalysisDatabase, LimitedCostTracker) -> (LimitedCostTracker, R) {
+
+//         let cost_tracker = LimitedCostTracker::new_free();
+//         let (_cost_tracker, result) = to_do(&mut self.store.as_analysis_db(), cost_tracker);
+//         result
+//     }
+// }
 
 pub fn with_memory_environment<F>(f: F, top_level: bool)
 where
@@ -106,25 +208,4 @@ where
 
     //         f(&mut owned_env)
     //     }
-}
-
-pub fn execute_on_network(s: &str, use_mainnet: bool) -> Value {
-    vm_execute_on_network(s, use_mainnet).unwrap().unwrap()
-}
-
-pub fn is_committed(v: &Value) -> bool {
-    eprintln!("is_committed?: {}", v);
-
-    match v {
-        Value::Response(ref data) => data.committed,
-        _ => false,
-    }
-}
-
-pub fn is_err_code(v: &Value, e: u128) -> bool {
-    eprintln!("is_err_code?: {}", v);
-    match v {
-        Value::Response(ref data) => !data.committed && *data.data == Value::UInt(e),
-        _ => false,
-    }
 }
